tests/pyproject.toml
[project]
name = "agent-home-system-tests"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
    "paho-mqtt>=2.1.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
py-modules = []

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
